+++ uninstall.sh
@@ -19,21 +19,19 @@ else
   WALLPAPER_DIR="$HOME/.local/share/wallpapers"
 fi
 
-SRC_DIR=$(cd $(dirname $0) && pwd)
-
 THEME_NAME=WhiteSur
 LATTE_DIR="$HOME/.config/latte"
 
 uninstall() {
   local name=${1}
 
-  [[ -d ${AURORAE_DIR}/${name} ]] && rm -rfv ${AURORAE_DIR}/${name}*
-  [[ -d ${PLASMA_DIR}/${name} ]] && rm -rfv ${PLASMA_DIR}/${name}*
-  [[ -f ${SCHEMES_DIR}/${name}.colors ]] && rm -rfv ${SCHEMES_DIR}/${name}*.colors
-  [[ -d ${LOOKFEEL_DIR}/com.github.vinceliuice.${name} ]] && rm -rfv ${LOOKFEEL_DIR}/com.github.vinceliuice.${name}*
-  [[ -d ${KVANTUM_DIR}/${name} ]] && rm -rfv ${KVANTUM_DIR}/${name}*
-  [[ -d ${WALLPAPER_DIR}/${name} ]] && rm -rfv ${WALLPAPER_DIR}/${name}
-  [[ -f ${LATTE_DIR}/${name}.layout.latte ]] && rm -rfv ${LATTE_DIR}/${name}.layout.latte
+  [[ -d ${AURORAE_DIR}/${name} ]] && rm -rfv "${AURORAE_DIR}"/"${name:?}"*
+  [[ -d ${PLASMA_DIR}/${name} ]] && rm -rfv "${PLASMA_DIR}"/"${name:?}"*
+  [[ -f ${SCHEMES_DIR}/${name}.colors ]] && rm -rfv "${SCHEMES_DIR}"/"${name:?}"*.colors
+  [[ -d ${LOOKFEEL_DIR}/com.github.vinceliuice.${name} ]] && rm -rfv "${LOOKFEEL_DIR}"/com.github.vinceliuice."${name:?}"*
+  [[ -d ${KVANTUM_DIR}/${name} ]] && rm -rfv "${KVANTUM_DIR}"/"${name:?}"*
+  [[ -d ${WALLPAPER_DIR}/${name} ]] && rm -rfv "${WALLPAPER_DIR}"/"${name:?}"
+  [[ -f ${LATTE_DIR}/${name}.layout.latte ]] && rm -rfv "${LATTE_DIR}"/"${name:?}".layout.latte
 }
 
 echo "Uninstalling '${THEME_NAME} kde themes'..."
